+++ core/src/main/java/com/novoda/noplayer/internal/listeners/PreparedListeners.java
@@ -1,4 +1,4 @@
-package com.novoda.noplayer.listeners;
+package com.novoda.noplayer.internal.listeners;
 
 import com.novoda.noplayer.Player;
 import com.novoda.noplayer.PlayerState;
